HW-5 UPDATE move package to adapter
